deprecate llama4, move apply_fsdp to common file
