+++ torchtitan/models/llama3/parallelize.py
@@ -7,16 +7,6 @@
 # This file applies the PT-D parallelisms (except pipeline parallelism) and various
 # training techniques (e.g. activation checkpointing and compile) to the Llama model.
 
-import torch
-import torch.nn as nn
-from torch.distributed.device_mesh import DeviceMesh
-from torch.distributed.fsdp import (
-    CPUOffloadPolicy,
-    DataParallelMeshDims,
-    fully_shard,
-    MixedPrecisionPolicy,
-)
-
 from torchtitan.config import (
     ActivationCheckpointConfig,
     CompileConfig,
@@ -28,15 +18,10 @@ from torchtitan.distributed import ParallelDims
 from torchtitan.distributed.activation_checkpoint import apply_ac
 from torchtitan.distributed.compile import apply_compile
 from torchtitan.distributed.context_parallel import apply_cp_to_forward
-from torchtitan.distributed.fsdp import (
-    disable_fsdp_gradient_division,
-    enable_fsdp_symm_mem,
-    get_fsdp_reshard_after_forward_policy,
-)
+from torchtitan.distributed.fsdp import apply_fsdp_to_decoder
 from torchtitan.distributed.full_dtensor import resolve_fsdp_mesh, validate_config
 from torchtitan.distributed.tensor_parallel import maybe_enable_async_tp
 from torchtitan.models.llama3.model import Llama3Model
-from torchtitan.tools.logging import logger
 
 
 def parallelize_llama(
@@ -94,8 +79,8 @@ def parallelize_llama(
     if model_compile_enabled:
         apply_compile(model, compile_config)
 
-    # Always run apply_fsdp -- with shard_degree=1 it is a no-op for the
-    # all-gather but still installs the MixedPrecisionPolicy.
+    # Always run apply_fsdp_to_decoder -- with shard_degree=1 it is a no-op for
+    # the all-gather but still installs the MixedPrecisionPolicy.
     if parallelism.spmd_backend == "full_dtensor":
         dp_mesh, dp_mesh_dims = resolve_fsdp_mesh(parallel_dims)
     else:
@@ -105,7 +90,7 @@ def parallelize_llama(
         dp_mesh = parallel_dims.get_mesh(names)
         dp_mesh_dims = None
 
-    apply_fsdp(
+    apply_fsdp_to_decoder(
         model,
         dp_mesh,
         param_dtype=TORCH_DTYPE_MAP[training.mixed_precision_param],
@@ -117,114 +102,4 @@ def parallelize_llama(
         enable_symm_mem=parallelism.enable_fsdp_symm_mem,
     )
 
-    if parallel_dims.dp_replicate_enabled:
-        logger.info("Applied HSDP to the model")
-    else:
-        logger.info("Applied FSDP to the model")
-
-    if training.enable_cpu_offload:
-        logger.info("Applied CPU Offloading to the model")
-
     return model
-
-
-def apply_fsdp(
-    model: nn.Module,
-    dp_mesh: DeviceMesh,
-    param_dtype: torch.dtype,
-    reduce_dtype: torch.dtype,
-    pp_enabled: bool,
-    cpu_offload: bool = False,
-    reshard_after_forward_policy: str = "default",
-    dp_mesh_dims: "DataParallelMeshDims | None" = None,
-    enable_symm_mem: bool = False,
-):
-    """
-    Apply data parallelism (via FSDP2) to the model.
-
-    Args:
-        model (nn.Module): The model to apply data parallelism to.
-        dp_mesh (DeviceMesh): The device mesh to use for data parallelism.
-        param_dtype (torch.dtype): The data type to use for model parameters.
-        reduce_dtype (torch.dtype): The data type to use for reduction operations.
-        pp_enabled (bool): Whether pipeline parallelism is enabled.
-        cpu_offload (bool, optional): Whether to offload model parameters to CPU. Defaults to False.
-        reshard_after_forward_policy (str, optional): The policy to use for resharding after forward pass. Defaults to "default".
-            Other options: "never", "always".
-            - "default" applies default resharding behavior, implementing "smart defaults" for known optimal scenarios.
-            - "always" will enable `reshard_after_forward` for all forward passes.
-            - "never" will disable `reshard_after_forward` for all forward passes.
-        dp_mesh_dims: Under full_dtensor, ``fully_shard`` must flatten
-            ``dp_shard`` and ``cp`` into a single FSDP shard dim, so it
-            needs to know which axes of the multi-D SPMD mesh are
-            data-parallel. We pass this explicitly via ``dp_mesh_dims``
-            rather than letting FSDP infer it from mesh axis names: the
-            naming contract between ``fully_shard`` and torchtitan is not
-            strong enough to infer safely, and an explicit declaration
-            avoids silent miscategorization when new mesh axes appear.
-        enable_symm_mem (bool): Whether to enable symmetric-memory FSDP
-            communication.
-    """
-    mp_policy = MixedPrecisionPolicy(
-        param_dtype=param_dtype,
-        reduce_dtype=reduce_dtype,
-        cast_forward_inputs=False,
-    )
-    fsdp_config = {"mesh": dp_mesh, "mp_policy": mp_policy}
-    if dp_mesh_dims is not None:
-        # pyrefly: ignore[bad-typed-dict-key]
-        fsdp_config["dp_mesh_dims"] = dp_mesh_dims
-    if cpu_offload:
-        # pyrefly: ignore[bad-typed-dict-key]
-        fsdp_config["offload_policy"] = CPUOffloadPolicy()
-
-    reshard_after_forward = get_fsdp_reshard_after_forward_policy(
-        reshard_after_forward_policy, pp_enabled
-    )
-
-    if getattr(model, "enable_weight_tying", False):
-        # When weights are tied, tok_embeddings and output share the same parameter.
-        # Group them together in one FSDP unit to avoid duplicate all-gathers.
-        modules = [
-            m
-            for m in (model.tok_embeddings, model.norm, model.lm_head)
-            if m is not None
-        ]
-        # pyrefly: ignore [no-matching-overload]
-        fully_shard(
-            modules,
-            **fsdp_config,
-            reshard_after_forward=reshard_after_forward_policy == "always",
-        )
-    else:
-        if model.tok_embeddings is not None:
-            # pyrefly: ignore [no-matching-overload]
-            fully_shard(
-                model.tok_embeddings,
-                **fsdp_config,
-                reshard_after_forward=reshard_after_forward,
-            )
-        # As an optimization, do not reshard_after_forward the last layers
-        # by default since FSDP would prefetch them immediately.
-        if model.norm is not None and model.lm_head is not None:
-            # pyrefly: ignore [no-matching-overload]
-            fully_shard(
-                [model.norm, model.lm_head],
-                **fsdp_config,
-                reshard_after_forward=reshard_after_forward_policy == "always",
-            )
-    # pyrefly: ignore [missing-attribute]
-    for layer_id, transformer_block in model.layers.items():
-        fully_shard(
-            transformer_block,
-            **fsdp_config,
-            reshard_after_forward=reshard_after_forward,
-        )
-
-    fully_shard(model, **fsdp_config)
-
-    if enable_symm_mem:
-        enable_fsdp_symm_mem(model)
-
-    # Disable FSDP's automatic gradient division for all FSDP modules
-    disable_fsdp_gradient_division(model)
