--- torchtitan/models/llama4/README.md
@@ -1,5 +1,0 @@
-#### Download Llama 4 tokenizer
-```bash
-# Llama 4 tokenizer.model
-python scripts/download_hf_assets.py --assets tokenizer --repo_id meta-llama/Llama-4-Scout-17B-16E --hf_token=...
-```
