--- torchtitan/models/llama4/config_registry.py
@@ -1,162 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-
-from torchtitan.components.checkpoint import CheckpointManager
-from torchtitan.components.loss import ChunkedCELoss
-from torchtitan.components.lr_scheduler import LRSchedulersContainer
-from torchtitan.components.metrics import MetricsProcessor
-from torchtitan.components.optimizer import default_adamw
-from torchtitan.components.quantization import Float8GroupedExpertsConverter
-from torchtitan.config import (
-    ActivationCheckpointConfig,
-    CompileConfig,
-    ParallelismConfig,
-    TrainingConfig,
-)
-from torchtitan.hf_datasets.text_datasets import HuggingFaceTextDataLoader
-from torchtitan.trainer import Trainer
-
-from . import model_registry
-
-
-def llama4_debugmodel() -> Trainer.Config:
-    return Trainer.Config(
-        loss=ChunkedCELoss.Config(),
-        hf_assets_path="./tests/assets/tokenizer",
-        metrics=MetricsProcessor.Config(log_freq=1),
-        model_spec=model_registry("debugmodel"),
-        dataloader=HuggingFaceTextDataLoader.Config(
-            dataset="c4_test",
-        ),
-        optimizer=default_adamw(lr=4e-3, eps=1e-15),
-        lr_scheduler=LRSchedulersContainer.Config(
-            warmup_steps=2,
-            decay_ratio=0.8,
-            decay_type="linear",
-            min_lr_factor=0.1,
-        ),
-        training=TrainingConfig(
-            local_batch_size=8,
-            seq_len=2048,
-            steps=10,
-        ),
-        parallelism=ParallelismConfig(
-            expert_parallel_degree=1,
-        ),
-        checkpoint=CheckpointManager.Config(
-            interval=10,
-            last_save_model_only=False,
-        ),
-        activation_checkpoint=ActivationCheckpointConfig(
-            mode="selective",
-        ),
-    )
-
-
-def llama4_debugmodel_ep() -> Trainer.Config:
-    config = llama4_debugmodel()
-    config.model_spec = model_registry("debugmodel")
-    return config
-
-
-def llama4_debugmodel_fp8() -> Trainer.Config:
-    compile_config = CompileConfig(enable=False, components=["model", "loss"])
-    return Trainer.Config(
-        loss=ChunkedCELoss.Config(),
-        hf_assets_path="./tests/assets/tokenizer",
-        metrics=MetricsProcessor.Config(log_freq=1),
-        model_spec=model_registry(
-            "debugmodel",
-            converters=[
-                Float8GroupedExpertsConverter.Config(
-                    model_compile_enabled=(
-                        compile_config.enable and "model" in compile_config.components
-                    ),
-                ),
-            ],
-        ),
-        dataloader=HuggingFaceTextDataLoader.Config(
-            dataset="c4_test",
-        ),
-        optimizer=default_adamw(lr=4e-3, eps=1e-15),
-        lr_scheduler=LRSchedulersContainer.Config(
-            warmup_steps=2,
-            decay_ratio=0.8,
-            decay_type="linear",
-            min_lr_factor=0.1,
-        ),
-        training=TrainingConfig(
-            local_batch_size=8,
-            seq_len=2048,
-            steps=10,
-        ),
-        parallelism=ParallelismConfig(
-            expert_parallel_degree=2,
-        ),
-        checkpoint=CheckpointManager.Config(
-            interval=10,
-            last_save_model_only=False,
-        ),
-        activation_checkpoint=ActivationCheckpointConfig(
-            mode="selective",
-        ),
-        compile=compile_config,
-    )
-
-
-def llama4_17bx128e() -> Trainer.Config:
-    return Trainer.Config(
-        loss=ChunkedCELoss.Config(),
-        hf_assets_path="./assets/hf/Llama-4-Maverick-17B-128E",
-        model_spec=model_registry("17bx128e"),
-        dataloader=HuggingFaceTextDataLoader.Config(
-            dataset="c4",
-        ),
-        optimizer=default_adamw(lr=4e-3, eps=1e-15),
-        lr_scheduler=LRSchedulersContainer.Config(
-            warmup_steps=600,
-            min_lr_factor=0.1,
-        ),
-        training=TrainingConfig(
-            local_batch_size=1,
-            seq_len=8192,
-            steps=3000,
-        ),
-        parallelism=ParallelismConfig(
-            tensor_parallel_degree=8,
-            pipeline_parallel_degree=4,
-            expert_parallel_degree=1,
-        ),
-        checkpoint=CheckpointManager.Config(interval=500),
-        activation_checkpoint=ActivationCheckpointConfig(mode="full"),
-    )
-
-
-def llama4_17bx16e() -> Trainer.Config:
-    return Trainer.Config(
-        loss=ChunkedCELoss.Config(),
-        hf_assets_path="./assets/hf/Llama-4-Scout-17B-16E",
-        model_spec=model_registry("17bx16e"),
-        dataloader=HuggingFaceTextDataLoader.Config(
-            dataset="c4",
-        ),
-        optimizer=default_adamw(lr=4e-3, eps=1e-15),
-        lr_scheduler=LRSchedulersContainer.Config(
-            warmup_steps=600,
-            min_lr_factor=0.1,
-        ),
-        training=TrainingConfig(
-            local_batch_size=8,
-            seq_len=8192,
-            steps=3000,
-        ),
-        parallelism=ParallelismConfig(
-            tensor_parallel_degree=8,
-            expert_parallel_degree=1,
-        ),
-        checkpoint=CheckpointManager.Config(interval=500),
-        activation_checkpoint=ActivationCheckpointConfig(mode="full"),
-    )
