+++ torchtitan/protocols/state_dict_adapter.py
@@ -114,21 +114,15 @@ class StateDictAdapter(BaseStateDictAdapter):
     def _validate_hf_rope_config(
         self,
         expected_rope_cls: type,
-        *,
-        allow_none: bool = False,
     ) -> None:
-        # Llama4 mixes RoPE and NoPE attention layers, so HF loading allows
-        # ``rope=None`` there. TODO: remove ``allow_none`` once Llama4 is deleted.
         for layer in self.model_config.layers:  # pyrefly: ignore [missing-attribute]
             rope = layer.attention.rope
-            if rope is None and allow_none:
-                continue
             if not isinstance(rope, expected_rope_cls):
                 expected_name = expected_rope_cls.__qualname__
-                message = f"HF checkpoint conversion assumes {expected_name}"
-                if allow_none:
-                    message += " for RoPE layers"
-                raise ValueError(f"{message}; got {type(rope).__name__}.")
+                raise ValueError(
+                    f"HF checkpoint conversion assumes {expected_name}; "
+                    f"got {type(rope).__name__}."
+                )
 
     def get_hf_storage_reader(
         self, path: str, from_quantized: bool = False
